Add Drawer Navigation to Showcase
