+++ demos/showcase/index.js
@@ -1,9 +1,9 @@
 /**
  * @format
  */
-
+import 'react-native-gesture-handler';
 import { AppRegistry } from 'react-native';
-import { App } from './App';
 import { name as appName } from './app.json';
+import {MainRouter} from "./router";
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(appName, () => MainRouter);
