+++ demos/showcase/package.json
@@ -18,10 +18,17 @@
         "@pxblue/icons-svg": "^1.0.18",
         "@pxblue/react-native-components": "^2.0.2",
         "@pxblue/react-native-themes": "^5.0.0-beta.0",
+        "@react-native-community/masked-view": "^0.1.10",
+        "@react-navigation/drawer": "^5.8.1",
+        "@react-navigation/native": "^5.5.0",
         "react": "16.11.0",
         "react-native": "0.62.2",
         "react-native-elements": "^2.0.0",
+        "react-native-gesture-handler": "^1.6.1",
         "react-native-paper": "^3.10.1",
+        "react-native-reanimated": "^1.9.0",
+        "react-native-safe-area-context": "^2.0.3",
+        "react-native-screens": "^2.8.0",
         "react-native-svg": "^12.1.0",
         "react-native-vector-icons": "^6.6.0"
     },
